commit e48c84d79572228e85cd82f9a15f6484ef9ae712
    10150 smbd needs smatch fixes
commit 2bc92914a1f7777163663081b69a9e5efa54c7d8
    10149 make latencytop smatch clean
commit 42bf02ad8ad8beded7a8046754d7aa9cf04b4deb
    10148 hald should skip smatch checks
commit 3a972b223858c0e81a04c85427cc2480979557ae
    10147 pyzfs should disable function return checks
commit 1c16f352c4ff31d252682640cf57c70463dd005f
    7173 sparc: __floatundidf undefined for sparcv9
commit 3ad945325009cea0c39efe3d207e0522c4eb2153
    10327 system-test-utiltest.mf has duplicate lines
commit b1659ed92cb99723e24e3144b2cc0be8663442d7
    10300 fem: cast between incompatible function types
commit ecd7dc692599292c0012e55e5c7ca9dffca4d1a1
    10299 kcpc.c: cast between incompatible function types
commit ebf373523fbbfcc7f9f0b7e35f0e82fac5492cfe
    10298 pcmcia: cast between incompatible function types
commit 3ebafc60580cf8a412e554e4babd8c3a0d3e98ea
    10297 cyclic_mp_init: cast between incompatible function types
commit 1fb83a8f4f86641886ceec86db9bd430ef9b046b
    10296 ficl: unused parameter vm in ficlDebugPrompt
commit cf9a187ca73bc899270af62ab1cd99f2e205cf41
    10293 usr/cmd/sgs - remove lint and lint libraries
commit 29a77b73a1e7f5d2b94fed029b676b4eba33440b
    10235 uts: boot needs simple tem to support mdb
commit c787c4e4eefec8e93a6441f70c306064870d781d
    10222 t_bind(3NSL) could mention data types for transports
commit 299e09dec5d1f56c12f76421d18079871b467bfc
    10051 cw(1) shouldn't have to translate -g
commit 096c97d62be876a03a0a8cdb0a540e9c84ec509f
    10267 ld and GCC disagree about i386 local dynamic TLS
commit 88e61e85cc7078a2b481258cf9441eeee7ff548d
    9899 cw(1onbld) should shadow more compilation

	Conflicts:
	usr/src/cmd/sgs/yacc/Makefile.targ
	usr/src/cmd/sgs/lex/Makefile.targ

+++ usr/src/uts/common/pcmcia/nexus/pcmcia.c
@@ -384,7 +384,7 @@ pcmcia_attach(dev_info_t *dip, anp_t *adapter)
 	mutex_enter(&pcmcia_enum_lock);
 	mutex_enter(&pcmcia_global_lock);
 	if (pcmcia_num_adapters == 0) {
-		pcmcia_cis_parser = (f_tt *)CISParser;
+		pcmcia_cis_parser = (f_tt *)(uintptr_t)CISParser;
 		cis_parser = (void *(*)(int, ...)) CISParser;
 		pcmcia_cs_event = (f_tt *)cs_event;
 		cs_socket_services = SocketServices;
@@ -1744,7 +1744,7 @@ pcmcia_set_em_handler(int (*handler)(), caddr_t events, int elen,
 		if (mif == NULL)
 			return (ENOSPC);
 
-		mif->mif_function = (void (*)())handler;
+		mif->mif_function = (void (*)())(uintptr_t)handler;
 		bcopy(events, mif->mif_events, elen);
 		mif->mif_id = id;
 		mutex_enter(&pcmcia_global_lock);
@@ -3815,7 +3815,7 @@ SSSetIRQHandler(set_irq_handler_t *handler)
 
 	retval = ddi_add_intr(dip, 0, handler->iblk_cookie,
 	    handler->idev_cookie,
-	    (uint32_t(*)(caddr_t)) handler->handler,
+	    (uint32_t(*)(caddr_t))(uintptr_t) handler->handler,
 	    handler->arg1);
 
 	if (retval == DDI_SUCCESS) {
@@ -4530,18 +4530,18 @@ is_subtractv(dev_info_t *dip)
 
 /*
  * pcmcia_pci_alloc()
- * 	allocate mem or I/O resource from the ancestor of the cardbus bridge.
- * 	First start from the parent node. If the parent is a subtractive
- * 	decode bridge and it does not have the requested resource, go up the
- * 	device tree to find the resource.
+ *	allocate mem or I/O resource from the ancestor of the cardbus bridge.
+ *	First start from the parent node. If the parent is a subtractive
+ *	decode bridge and it does not have the requested resource, go up the
+ *	device tree to find the resource.
  *
- * 	dip		the parent node of the cardbus bridge
+ *	dip		the parent node of the cardbus bridge
  *
- * 	res_dip		returns a pointer to the node from which the
- * 			resource is obtained. *res_dip could point to
- * 			the parent or a higher level ancestor. *res_dip
- * 			should be saved by the caller and later passed
- * 			to pcmcia_ra_free();
+ *	res_dip		returns a pointer to the node from which the
+ *			resource is obtained. *res_dip could point to
+ *			the parent or a higher level ancestor. *res_dip
+ *			should be saved by the caller and later passed
+ *			to pcmcia_ra_free();
  */
 int
 pcmcia_pci_alloc(dev_info_t *dip, ndi_ra_request_t *req, ra_return_t *ret,
@@ -5292,7 +5292,7 @@ pcmcia_intr_enable_isr(dev_info_t *dip, dev_info_t *rdip,
 
 	handler.socket = sockp->ls_socket;
 	handler.irq = irq;
-	handler.handler = (f_tt *)hdlp->ih_cb_func;
+	handler.handler = (f_tt *)(uintptr_t)hdlp->ih_cb_func;
 	handler.arg1 = hdlp->ih_cb_arg1;
 	handler.arg2 = hdlp->ih_cb_arg2;
 	handler.handler_id = (uint32_t)(uintptr_t)rdip;
